Render rich text on the map page
Until now, only elixir could transform the raw output of draft-js to
HTML. Now it can also be done with the `<RichText />` component.

To keep the ugly bits in one place, I added a new module RichText, which
defines some structs that can either be consumed by a Phoenix view, or
by a react component. The translation of the draft-js format only exists
in the Elixir module.

This looks like a lot of code for a very small use case, might want to
see if it can be done in a simpler, terser way.