[project][show] add video modal
Using micromodal.js seems like a good idea, it makes the modal
accessible, and seems to be pretty lightweight.

There's already a bug when closing the modal using "ESC" though, the
iframe does not appear if you reopen the modal. But the overlay is here,
as is the html document of the iframe, so it looks like it's an iframe
problem (both youtube and vimeo players do this).