Do not use AJAX for navigation
The *only* thing that made us use AJAX for navigation was the page
transitions and logo animation.

It turns out that the animation can simply be launched when a link a
clicked, and letting the browser do a full reload. The browser is smart
enough to not make the UI flash.

The main advantages of this method is that it's must less code, and full
page reloads are nice when working with javascript. Also, we get scroll
position management when browsing back pages for free.